[JSON][WL] update email, numeric, text fields - rename uiType email as email-field - rename uiType numeric as numeric-field - rename uiType text as text-field - update interfaces, stories and file names to new name
