+++ src/components/fields/text-field/index.ts
@@ -1,0 +1,2 @@
+export * from "./text-field";
+export * from "./types";
