--- src/components/fields/textfield/index.ts
@@ -1,2 +1,0 @@
-export * from "./textfield";
-export * from "./types";
