+++ src/components/frontend-engine/types.ts
@@ -6,14 +6,14 @@ import {
 	IChipsSchema,
 	IContactNumberSchema,
 	IDateInputSchema,
-	IEmailSchema,
+	IEmailFieldSchema,
 	IMultiSelectSchema,
-	INumberSchema,
+	INumericFieldSchema,
 	IRadioButtonGroupSchema,
 	ISelectSchema,
 	ISubmitButtonSchema,
+	ITextFieldSchema,
 	ITextareaSchema,
-	ITextfieldSchema,
 	ITimePickerSchema,
 } from "../fields";
 import { IYupValidationRule, TRenderRules, TYupSchemaType } from "./yup";
@@ -54,9 +54,9 @@ export interface IFrontendEngineData<V = undefined> {
 
 export type TFrontendEngineFieldSchema<V = undefined> =
 	| ITextareaSchema<V>
-	| ITextfieldSchema<V>
-	| IEmailSchema<V>
-	| INumberSchema<V>
+	| ITextFieldSchema<V>
+	| IEmailFieldSchema<V>
+	| INumericFieldSchema<V>
 	| ISubmitButtonSchema
 	| ISelectSchema<V>
 	| IMultiSelectSchema<V>
@@ -141,9 +141,9 @@ export type TComponentOmitProps<T, V = undefined> = Omit<T, UnionOptionalKeys<V>
 
 export enum EFieldType {
 	TEXTAREA = "Textarea",
-	TEXT = "TextField",
-	NUMERIC = "TextField",
-	EMAIL = "TextField",
+	"TEXT-FIELD" = "TextField",
+	"NUMERIC-FIELD" = "TextField",
+	"EMAIL-FIELD" = "TextField",
 	SUBMIT = "SubmitButton",
 	SELECT = "Select",
 	"MULTI-SELECT" = "MultiSelect",
