+++ src/stories/2-frontend-engine/conditional-rendering/conditional-rendering-array.stories.tsx
@@ -48,7 +48,7 @@ Length.args = {
 	},
 	field2: {
 		label: "Field 2",
-		uiType: "text",
+		uiType: "text-field",
 		showIf: [{ field1: [{ filled: true }, { length: 2 }] }],
 		validation: [{ required: true }],
 	},
@@ -72,7 +72,7 @@ MinItems.args = {
 	},
 	field2: {
 		label: "Field 2",
-		uiType: "text",
+		uiType: "text-field",
 		showIf: [{ field1: [{ filled: true }, { min: 2 }] }],
 		validation: [{ required: true }],
 	},
@@ -96,7 +96,7 @@ MaxItems.args = {
 	},
 	field2: {
 		label: "Field 2",
-		uiType: "text",
+		uiType: "text-field",
 		showIf: [{ field1: [{ filled: true }, { max: 2 }] }],
 		validation: [{ required: true }],
 	},
@@ -120,7 +120,7 @@ Includes.args = {
 	},
 	field2: {
 		label: "Field 2",
-		uiType: "text",
+		uiType: "text-field",
 		showIf: [{ field1: [{ filled: true }, { includes: ["Apple", "Berry"] }] }],
 		validation: [{ required: true }],
 	},
@@ -144,7 +144,7 @@ Excludes.args = {
 	},
 	field2: {
 		label: "Field 2",
-		uiType: "text",
+		uiType: "text-field",
 		showIf: [{ field1: [{ filled: true }, { excludes: ["Apple", "Berry"] }] }],
 		validation: [{ required: true }],
 	},
